Move phpunit to "require-dev".
PHPUnit is not needed by the runtime library. It is only needed to run the
unit tests when developing the library. This change allows phpunit to be
installed by composer when `composer install --dev` but keeps it from being
installed when a user is including sphpdox in their own project's
composer.json file.